requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
fastapi==0.117.1
filelock==3.19.1
fonttools==4.60.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
narwhals==2.5.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
polars==1.33.1
protobuf==6.32.1
psutil==7.1.0
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scikit-image==0.25.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.48.0
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.9.20
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.203
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.37.0
